refactor: 프론트 ui 및 백 수동 배치 로직 위치 수정
